+++ docs/Changelog.md
@@ -1,5 +1,5 @@
 
-5.4.9.2 Beta (2021-??-??)
+5.4.9.2 Beta (2021-08-08)
 
 - Manual translated to simplified Chinese. (hooke007)
 - watch-later-options support added to conf editor. (hooke007)
@@ -15,6 +15,9 @@
 - Some scroll bars where replaced with Windows 10 styled scroll bars,
   complex code used from HandyControl project.
 - Some UI elements use rounded corners.
+- The recent list can also be shown in the command palette:
+  Alt+r script-message mpv.net show-recent #menu: View > Show Recent
+- The recent context menu removes the folder info in case of very long paths.
 - libmpv shinchiro 2021-08-01